utility/file-ops: Return size_t, not int, from read_file_data().

+++ utility/file-ops.cpp
@@ -21,7 +21,7 @@ using std::string;
 
 namespace oodles {
 
-int read_file_data(const string &path, string &output)
+size_t read_file_data(const string &path, string &output)
 {
     ifstream input(path.c_str());
